message routes refactored

+++ app/routes/message_routes.py
@@ -12,11 +12,11 @@ message_bp = Blueprint('messages', __name__)
 def new_message():
     form = MessageForm()
     form['csrf_token'].data = request.cookies['csrf_token']
-    
+
     if form.validate_on_submit():
         new_message = Message()
         form.populate_obj(new_message)
-        
+
         db.session.add(new_message)
         db.session.commit()
         return new_message.to_dict()
@@ -25,29 +25,30 @@ def new_message():
         return form.errors
 
 
-@message_bp.route('/<int:id>', methods=['GET', 'DELETE'])
-def delete_message(id):
+@message_bp.route('/<int:id>', methods=['GET', 'DELETE', 'PUT'])
+def get_message_by_id(id):
     message = Message.query.get(id)
-    
-    if request.method == 'DELETE':
-        db.session.delete(message)
-        db.session.commit()
-        return {'message': 'Message Deleted!'}
 
-    return message.to_dict()
+    if message:
 
+        if request.method == 'GET':
+            message_dict = message.to_dict()
+            message_dict['author'] = message.author.to_dict()
+            return message_dict
 
-@message_bp.route('/<int:id>', methods=["PUT"])
-def update_message(id):
-    form = MessageForm()
-    form['csrf_token'].data = request.cookies['csrf_token']
-    message_update = Message.query.get(id)
-    
-    if request.method == "PUT":
-        if form.validate_on_submit():
-            message_update.body = form.data['body']
+        if request.method == 'DELETE':
+            db.session.delete(message)
             db.session.commit()
-        
-            return message_update.to_dict()
-    
-        return form.errors
+            return {'message': 'Message Deleted!'}
+
+        if request.method == 'PUT':
+            form = MessageForm()
+            form['csrf_token'].data = request.cookies['csrf_token']
+            if form.validate_on_submit():
+                message.body = form.data['body']
+                db.session.commit()
+                return message.to_dict()
+
+            return form.errors
+
+    return { "error": "Message not found", "errorCode" : 404 }, 404
